apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
pytesseract==0.3.10
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
